app_domains/requirements_v2.txt
aiodns==2.0.0
aiohttp==3.6.2
args==0.1.0
async-timeout==3.0.1
asyncio==3.4.3
asyncpool==1.0
attrs==19.3.0
beautifulsoup4==4.9.0
blis==0.4.1
Bottleneck==1.3.2
brotlipy==0.7.0
catalogue==1.0.0
cchardet==2.1.6
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
clint==0.5.1
cltk==0.1.117
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
dill==0.3.1.1
dnspython==1.16.0
et-xmlfile==1.0.1
fastai==1.0.57
fastprogress==0.2.3
gitdb==4.0.4
GitPython==3.1.1
hazm==0.7.0
hebrew-tokenizer==1.0.3
idna==2.9
importlib-metadata==1.6.0
inflect==0.2.5
inltk==0.8.1
jdcal==1.4.1
jieba==0.42.1
joblib==0.14.1
JPype1==0.7.3
kiwisolver==1.2.0
konlpy==0.5.2
langdetect==1.0.8
languageflow==1.1.13a1
libwapiti==0.2.1
lxml==4.5.0
matplotlib==3.2.1
multidict==4.7.5
murmurhash==1.0.2
nltk==3.5
numexpr==2.7.1
numpy==1.18.3
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
openpyxl==3.0.3
packaging==20.3
pandas==1.0.3
Pillow==7.1.1
plac==1.1.3
preshed==3.0.2
psycopg2==2.8.5
pycares==3.1.1
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pythainlp==2.1.4
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
pyuca==1.2
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-file==1.4.3
requests-oauthlib==1.3.0
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
sentencepiece==0.1.85
six==1.14.0
smmap==3.0.2
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
tabulate==0.8.7
thinc==7.4.0
tinydb==3.15.2
tinysegmenter==0.4
tldextract==2.2.2
torch==1.3.1+cpu
torchvision==0.6.0
tqdm==4.45.0
tweepy==3.8.0
typing==3.7.4.1
underthesea==1.1.17
Unipath==1.1
urllib3==1.25.9
uvloop==0.14.0
wasabi==0.6.0
Whoosh==2.7.4
word-forms==1.0
xgboost==1.0.2
yarl==1.4.2
zipp==3.1.0
